support file templates per project (merge from idea)

+++ platform/lang-impl/src/com/intellij/ide/fileTemplates/impl/FTManager.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2000-2011 JetBrains s.r.o.
+ * Copyright 2000-2015 JetBrains s.r.o.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,13 +15,11 @@
  */
 package com.intellij.ide.fileTemplates.impl;
 
-import com.intellij.CommonBundle;
-import com.intellij.ide.IdeBundle;
 import com.intellij.ide.fileTemplates.FileTemplate;
-import com.intellij.openapi.application.PathManager;
 import com.intellij.openapi.diagnostic.Logger;
+import com.intellij.openapi.fileTypes.FileTypeManager;
 import com.intellij.openapi.project.ex.ProjectManagerEx;
-import com.intellij.openapi.ui.Messages;
+import com.intellij.openapi.util.Pair;
 import com.intellij.openapi.util.io.FileUtil;
 import com.intellij.openapi.util.text.StringUtil;
 import com.intellij.openapi.vfs.CharsetToolkit;
@@ -39,26 +37,37 @@ import java.util.*;
  */
 class FTManager {
   private static final Logger LOG = Logger.getInstance("#com.intellij.ide.fileTemplates.impl.FTManager");
-  public static final String TEMPLATES_DIR = "fileTemplates";
   public static final String DEFAULT_TEMPLATE_EXTENSION = "ft";
   public static final String TEMPLATE_EXTENSION_SUFFIX = "." + DEFAULT_TEMPLATE_EXTENSION;
-  public static final String CONTENT_ENCODING = CharsetToolkit.UTF8;
+  private static final String ENCODED_NAME_EXT_DELIMITER = "\u0F0Fext\u0F0F.";
 
   private final String myName;
   private final boolean myInternal;
-  private final String myTemplatesDir;
-  private final Map<String, FileTemplateBase> myTemplates = new HashMap<String, FileTemplateBase>();
+  private final File myTemplatesDir;
+  @Nullable
+  private final FTManager myOriginal;
+  private final Map<String, FileTemplateBase> myTemplates = new HashMap<>();
   private volatile List<FileTemplateBase> mySortedTemplates;
-  private final List<DefaultTemplate> myDefaultTemplates = new ArrayList<DefaultTemplate>();
+  private final List<DefaultTemplate> myDefaultTemplates = new ArrayList<>();
 
-  FTManager(@NotNull @NonNls String name, @NotNull @NonNls String defaultTemplatesDirName) {
+  FTManager(@NotNull @NonNls String name, @NotNull @NonNls File defaultTemplatesDirName) {
     this(name, defaultTemplatesDirName, false);
   }
 
-  FTManager(@NotNull @NonNls String name, @NotNull @NonNls String defaultTemplatesDirName, boolean internal) {
+  FTManager(@NotNull @NonNls String name, @NotNull @NonNls File defaultTemplatesDirName, boolean internal) {
     myName = name;
     myInternal = internal;
-    myTemplatesDir = TEMPLATES_DIR + (defaultTemplatesDirName.equals(".") ? "" : File.separator + defaultTemplatesDirName);
+    myTemplatesDir = defaultTemplatesDirName;
+    myOriginal = null;
+  }
+
+  FTManager(@NotNull FTManager original) {
+    myOriginal = original;
+    myName = original.getName();
+    myTemplatesDir = original.myTemplatesDir;
+    myInternal = original.myInternal;
+    myTemplates.putAll(original.myTemplates);
+    myDefaultTemplates.addAll(original.myDefaultTemplates);
   }
 
   public String getName() {
@@ -69,13 +78,8 @@ class FTManager {
   public Collection<FileTemplateBase> getAllTemplates(boolean includeDisabled) {
     List<FileTemplateBase> sorted = mySortedTemplates;
     if (sorted == null) {
-      sorted = new ArrayList<FileTemplateBase>(myTemplates.values());
-      Collections.sort(sorted, new Comparator<FileTemplateBase>() {
-        @Override
-        public int compare(FileTemplateBase t1, FileTemplateBase t2) {
-          return t1.getName().compareToIgnoreCase(t2.getName());
-        }
-      });
+      sorted = new ArrayList<>(getTemplates().values());
+      Collections.sort(sorted, (t1, t2) -> t1.getName().compareToIgnoreCase(t2.getName()));
       mySortedTemplates = sorted;
     }
 
@@ -83,7 +87,7 @@ class FTManager {
       return Collections.unmodifiableCollection(sorted);
     }
 
-    final List<FileTemplateBase> list = new ArrayList<FileTemplateBase>(sorted.size());
+    final List<FileTemplateBase> list = new ArrayList<>(sorted.size());
     for (FileTemplateBase template : sorted) {
       if (template instanceof BundledFileTemplate && !((BundledFileTemplate)template).isEnabled()) {
         continue;
@@ -99,7 +103,7 @@ class FTManager {
    */
   @Nullable
   public FileTemplateBase getTemplate(@NotNull String templateQname) {
-    return myTemplates.get(templateQname);
+    return getTemplates().get(templateQname);
   }
 
   /**
@@ -109,7 +113,7 @@ class FTManager {
    */
   @Nullable
   public FileTemplateBase findTemplateByName(@NotNull String templateName) {
-    final FileTemplateBase template = myTemplates.get(templateName);
+    final FileTemplateBase template = getTemplates().get(templateName);
     if (template != null) {
       final boolean isEnabled = !(template instanceof BundledFileTemplate) || ((BundledFileTemplate)template).isEnabled();
       if (isEnabled) {
@@ -119,8 +123,11 @@ class FTManager {
     // templateName must be non-qualified name, since previous lookup found nothing
     for (FileTemplateBase t : getAllTemplates(false)) {
       final String qName = t.getQualifiedName();
-      if (qName.startsWith(templateName) && qName.charAt(templateName.length()) == '.') {
-        return t;
+      if (qName.startsWith(templateName) && qName.length() > templateName.length()) {
+        String remainder = qName.substring(templateName.length());
+        if (remainder.startsWith(ENCODED_NAME_EXT_DELIMITER) || remainder.charAt(0) == '.') {
+          return t;
+        }
       }
     }
     return null;
@@ -132,7 +139,7 @@ class FTManager {
     FileTemplateBase template = getTemplate(qName);
     if (template == null) {
       template = new CustomFileTemplate(name, extension);
-      myTemplates.put(qName, template);
+      getTemplates().put(qName, template);
       mySortedTemplates = null;
     }
     else {
@@ -144,9 +151,9 @@ class FTManager {
   }
 
   public void removeTemplate(@NotNull String qName) {
-    final FileTemplateBase template = myTemplates.get(qName);
+    final FileTemplateBase template = getTemplates().get(qName);
     if (template instanceof CustomFileTemplate) {
-      myTemplates.remove(qName);
+      getTemplates().remove(qName);
       mySortedTemplates = null;
     }
     else if (template instanceof BundledFileTemplate){
@@ -154,26 +161,32 @@ class FTManager {
     }
   }
 
-  public void updateTemplates(Collection<FileTemplate> newTemplates) {
-    final Set<String> toDisable = new HashSet<String>();
+  public void updateTemplates(@NotNull Collection<FileTemplate> newTemplates) {
+    final Set<String> toDisable = new HashSet<>();
     for (DefaultTemplate template : myDefaultTemplates) {
       toDisable.add(template.getQualifiedName());
     }
     for (FileTemplate template : newTemplates) {
       toDisable.remove(((FileTemplateBase)template).getQualifiedName());
     }
-    myTemplates.clear();
+    restoreDefaults(toDisable);
+    for (FileTemplate template : newTemplates) {
+      final FileTemplateBase _template = addTemplate(template.getName(), template.getExtension());
+      _template.setText(template.getText());
+      _template.setReformatCode(template.isReformatCode());
+      _template.setLiveTemplateEnabled(template.isLiveTemplateEnabled());
+    }
+    saveTemplates(true);
+  }
+
+  private void restoreDefaults(Set<String> toDisable) {
+    getTemplates().clear();
     mySortedTemplates = null;
     for (DefaultTemplate template : myDefaultTemplates) {
       final BundledFileTemplate bundled = createAndStoreBundledTemplate(template);
       if (toDisable.contains(bundled.getQualifiedName())) {
         bundled.setEnabled(false);
       }
-    }
-    for (FileTemplate template : newTemplates) {
-      final FileTemplateBase _template = addTemplate(template.getName(), template.getExtension());
-      _template.setText(template.getText());
-      _template.setReformatCode(template.isReformatCode());
     }
   }
 
@@ -185,20 +198,76 @@ class FTManager {
   private BundledFileTemplate createAndStoreBundledTemplate(DefaultTemplate template) {
     final BundledFileTemplate bundled = new BundledFileTemplate(template, myInternal);
     final String qName = bundled.getQualifiedName();
-    final FileTemplateBase previous = myTemplates.put(qName, bundled);
+    final FileTemplateBase previous = getTemplates().put(qName, bundled);
     mySortedTemplates = null;
 
-    LOG.assertTrue(previous == null, "Duplicate bundled template " + qName);
+    LOG.assertTrue(previous == null, "Duplicate bundled template " + qName +
+                                     " [" + template.getTemplateURL() + ", " + previous + ']');
     return bundled;
   }
 
-  void saveTemplates() {
+  void loadCustomizedContent() {
+    final File configRoot = getConfigRoot(false);
+    final File[] configFiles = configRoot.listFiles();
+    if (configFiles == null) {
+      return;
+    }
+
+    final List<File> templateWithDefaultExtension = new ArrayList<>();
+    final Set<String> processedNames = new HashSet<>();
+
+    for (File file : configFiles) {
+      if (file.isDirectory() || FileTypeManager.getInstance().isFileIgnored(file.getName()) || file.isHidden()) {
+        continue;
+      }
+      final String name = file.getName();
+      if (name.endsWith(TEMPLATE_EXTENSION_SUFFIX)) {
+        templateWithDefaultExtension.add(file);
+      }
+      else {
+        processedNames.add(name);
+        addTemplateFromFile(name, file);
+      }
+    }
+
+    for (File file : templateWithDefaultExtension) {
+      String name = file.getName();
+      // cut default template extension
+      name = name.substring(0, name.length() - TEMPLATE_EXTENSION_SUFFIX.length());
+      if (!processedNames.contains(name)) {
+        addTemplateFromFile(name, file);
+      }
+      FileUtil.delete(file);
+    }
+  }
+
+  private void addTemplateFromFile(String fileName, File file) {
+    Pair<String,String> nameExt = decodeFileName(fileName);
+    final String extension = nameExt.second;
+    final String templateQName = nameExt.first;
+    if (templateQName.length() == 0) {
+      return;
+    }
+    try {
+      final String text = FileUtil.loadFile(file, CharsetToolkit.UTF8_CHARSET);
+      addTemplate(templateQName, extension).setText(text);
+    }
+    catch (IOException e) {
+      LOG.error(e);
+    }
+  }
+
+  public void saveTemplates() {
+    saveTemplates(false);
+  }
+
+  private void saveTemplates(boolean removeDeleted) {
     final File configRoot = getConfigRoot(true);
 
     final File[] files = configRoot.listFiles();
 
-    final Set<String> allNames = new HashSet<String>();
-    final Map<String, File> templatesOnDisk = files != null && files.length > 0? new HashMap<String, File>() : Collections.<String, File>emptyMap();
+    final Set<String> allNames = new HashSet<>();
+    final Map<String, File> templatesOnDisk = files != null && files.length > 0 ? new HashMap<>() : Collections.<String, File>emptyMap();
     if (files != null) {
       for (File file : files) {
         if (!file.isDirectory()) {
@@ -209,7 +278,7 @@ class FTManager {
       }
     }
 
-    final Map<String, FileTemplateBase> templatesToSave = new HashMap<String, FileTemplateBase>();
+    final Map<String, FileTemplateBase> templatesToSave = new HashMap<>();
 
     for (FileTemplateBase template : getAllTemplates(true)) {
       if (template instanceof BundledFileTemplate && !((BundledFileTemplate)template).isTextModified()) {
@@ -236,12 +305,14 @@ class FTManager {
         }
         else if (templateToSave == null) {
           // template was removed
-          FileUtil.delete(customizedTemplateFile);
+          if (removeDeleted) {
+            FileUtil.delete(customizedTemplateFile);
+          }
         }
         else {
           // both customized content on disk and corresponding template are present
           try {
-            final String diskText = StringUtil.convertLineSeparators(FileUtil.loadFile(customizedTemplateFile, CONTENT_ENCODING));
+            final String diskText = StringUtil.convertLineSeparators(FileUtil.loadFile(customizedTemplateFile, CharsetToolkit.UTF8_CHARSET));
             final String templateText = templateToSave.getText();
             if (!diskText.equals(templateText)) {
               // save only if texts differ to avoid unnecessary file touching
@@ -261,7 +332,7 @@ class FTManager {
    *  todo: review saving algorithm
    */
   private static void saveTemplate(File parentDir, FileTemplateBase template, final String lineSeparator) throws IOException {
-    final File templateFile = new File(parentDir, template.getName() + "." + template.getExtension());
+    final File templateFile = new File(parentDir, encodeFileName(template.getName(), template.getExtension()));
 
     FileOutputStream fileOutputStream;
     try {
@@ -272,16 +343,7 @@ class FTManager {
       FileUtil.delete(templateFile);
       fileOutputStream = new FileOutputStream(templateFile);
     }
-    OutputStreamWriter outputStreamWriter;
-    try{
-      outputStreamWriter = new OutputStreamWriter(fileOutputStream, CONTENT_ENCODING);
-    }
-    catch (UnsupportedEncodingException e){
-      Messages.showMessageDialog(IdeBundle.message("error.unable.to.save.file.template.using.encoding", template.getName(),
-                                                   CONTENT_ENCODING),
-                                 CommonBundle.getErrorTitle(), Messages.getErrorIcon());
-      outputStreamWriter = new OutputStreamWriter(fileOutputStream);
-    }
+    OutputStreamWriter outputStreamWriter = new OutputStreamWriter(fileOutputStream, CharsetToolkit.UTF8_CHARSET);
     String content = template.getText();
 
     if (!lineSeparator.equals("\n")){
@@ -293,15 +355,12 @@ class FTManager {
     fileOutputStream.close();
   }
 
+  @NotNull
   public File getConfigRoot(boolean create) {
-    final File templatesPath = new File(PathManager.getConfigPath(), myTemplatesDir);
-    if (create) {
-      final boolean created = templatesPath.mkdirs();
-      if (!created && !templatesPath.exists()) {
-        LOG.info("Cannot create directory: " + templatesPath.getAbsolutePath());
-      }
+    if (create && !myTemplatesDir.mkdirs() && !myTemplatesDir.exists()) {
+      LOG.info("Cannot create directory: " + myTemplatesDir.getAbsolutePath());
     }
-    return templatesPath;
+    return myTemplatesDir;
   }
 
   @Override
@@ -309,4 +368,24 @@ class FTManager {
     return myName + " file template manager";
   }
 
+  public static String encodeFileName(String templateName, String extension) {
+    String nameExtDelimiter = extension.contains(".") ? ENCODED_NAME_EXT_DELIMITER : ".";
+    return templateName + nameExtDelimiter + extension;
+  }
+
+  public static Pair<String,String> decodeFileName(String fileName) {
+    String name = fileName;
+    String ext = "";
+    String nameExtDelimiter = fileName.contains(ENCODED_NAME_EXT_DELIMITER) ? ENCODED_NAME_EXT_DELIMITER : ".";
+    int extIndex = fileName.lastIndexOf(nameExtDelimiter);
+    if (extIndex >= 0) {
+      name = fileName.substring(0, extIndex);
+      ext = fileName.substring(extIndex + nameExtDelimiter.length());
+    }
+    return Pair.create(name, ext);
+  }
+
+  public Map<String, FileTemplateBase> getTemplates() {
+    return myOriginal != null ? myOriginal.myTemplates : myTemplates;
+  }
 }
